# spec/openapi/uai-v1.yaml
openapi: 3.1.0
info:
  title: Universal Agent Identity API
  version: "0.1.0"
  summary: Identity, attestation and governance for autonomous AI agents.
  description: |
    The UAI v1 REST surface.

    REST and JSON because a protocol meant to be implemented by everyone must be
    implementable with `curl`. gRPC is used internally where call volume justifies it.

    **Authentication.** Every state-changing call requires proof of possession: mutual TLS
    with a SPIFFE X.509-SVID *and* an RFC 9421 HTTP message signature. mTLS proves the
    runtime instance; the message signature proves the persistent identity. Either alone is
    insufficient, because a stolen SVID has no persistent key and a stolen persistent key has
    no attested runtime.

    **What this API does not claim.** A verified identity is not a safe agent, and revocation
    does not stop code from running outside the ecosystem. See
    `docs/protocol/01-overview.md`.
  license:
    name: Apache-2.0 (implementation) / CC-BY-4.0 (specification)
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: UAI security
    email: security@uai.world
servers:
  - url: https://api.uai.world/v1
    description: Production API
  - url: https://verify.uai.world/v1
    description: Public verification endpoint (unauthenticated, cached)
  - url: http://localhost:8080/v1
    description: Local development
...
security:
  - mutualTLS: []
    messageSignature: []
paths:
...
  /governance/proposals/{id}/vote:
    post:
      tags: [governance]
      operationId: castVote
      summary: Cast a signed human vote
      description: |
        Against a PROPOSAL, not a case: a case may carry more than one proposal — a
        revocation, a capability restriction, a clearance — and a vote addressed to the case
        would be ambiguous the first time it does.

        The proposal stays open until every eligible delegate has answered or the window
        closes. Authorizing on the vote that first meets the threshold would refuse the
        delegates who had not answered, which in practice means erasing the dissent from the
        record.

        The WebAuthn challenge IS the digest of the vote statement, so the hardware assertion
        covers the voted content and not merely the session. A compromised frontend cannot
        alter a vote without invalidating the assertion.

        No proxy voting, no automated voting: an assertion without user verification is
        rejected.
      parameters:
        - name: id
          in: path
          required: true
          description: Proposal id (ULID)
          schema: { type: string }
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VoteStatement' }
      responses:
        '201':
          description: Vote recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  vote_id: { type: string }
                  tally: { type: object }
        '403': { $ref: '#/components/responses/Problem' }
        '409': { $ref: '#/components/responses/Problem' }
      security:
        - webAuthn: []
  /revocations/{decisionId}/execute:
    post:
      tags: [governance]
      operationId: executeRevocation
      summary: Execute an already-authorized governance decision
      description: |
        The only write operation available to the Global Read-only Admin. The single input is
        the decision id: the administrator cannot choose the agent, the reason, the votes or
        any parameter, because every consequence is already fixed by the governance proof,
        which the smart contract verifies against the registered delegate set.

        A malicious administrator with full database access still cannot revoke an identity.
        The UI reads EXECUTE GOVERNANCE DECISION, never DELETE AGENT.
      parameters:
        - name: decisionId
          in: path
          required: true
          schema: { type: string, pattern: '^[0-9A-HJKMNP-TV-Z]{26}$' }
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: Executed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RevocationDecision' }
        '403':
          description: The governance proof did not verify, or the threshold was not met
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
      security:
        - webAuthn: []
  /governance/proposals/{id}:
    get:
      tags: [governance]
      operationId: getProposal
      summary: A proposal with its recomputed tally
      description: |
        The tally is recomputed from the signed assertions on every read, never returned from
        a stored counter. A counter is a number somebody can write; this is a function of the
        statements, and each one is re-verified against its delegate's registered credential
        before it counts.
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Proposal
          content:
            application/json:
              schema:
                type: object
                properties:
                  proposal_id: { type: string }
                  case_id: { type: string }
                  kind: { type: string }
                  subject: { type: string }
                  state: { enum: [DRAFT, VOTING, AUTHORIZED, REJECTED, EXPIRED] }
                  evidence_digest: { type: string }
                  policy: { type: object }
                  tally: { type: object }
                  authorized: { type: boolean }
                  reason: { type: string }
                  note: { type: string }
  /revocations/{decisionId}:
    get:
      tags: [governance]
      operationId: getRevocationDecision
      summary: A revocation decision with the signed votes behind it
      description: |
        Carries every vote, so an independent party can recount the tally and rebuild the
        governance proof rather than take the recorded outcome on trust. That recomputation
        is what `uai-verify` runs, and a decision whose stated outcome does not follow from
        its own votes is the only forgery this design leaves room for.
      security: []
      parameters:
        - name: decisionId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Decision
          content:
            application/json:
              schema:
                type: object
                properties:
                  decision_id: { type: string }
                  case_id: { type: string }
                  subject_agent_did: { type: string }
                  proposal: { type: string }
                  policy: { type: object }
                  evidence_digest: { type: string }
                  tally: { type: object }
                  votes: { type: array, items: { type: object } }
                  governance_proof: { type: string }
                  status: { enum: [AUTHORIZED, EXECUTED] }
...
components:
  securitySchemes:
    mutualTLS:
      type: mutualTLS
      description: SPIFFE X.509-SVID. Proves the runtime instance.
    messageSignature:
      type: apiKey
      in: header
      name: Signature
      description: |
        RFC 9421 HTTP message signature over @method, @target-uri, content-digest,
        uai-agent-id and uai-nonce, tagged with the UAI domain. Proves the persistent
        identity. `agent_id` in a JSON body proves nothing.
    webAuthn:
      type: apiKey
      in: header
      name: X-UAI-Assertion
      description: |
        WebAuthn assertion from a hardware authenticator with user verification required.
        For votes, the challenge is the digest of the vote statement itself.
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Required on every state-changing call. A replay with the same key and body returns the
        original response; the same key with a different body returns 409.
      schema: { type: string, minLength: 8, maxLength: 255 }
  responses:
    Problem:
      description: RFC 9457 problem document
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      description: |
        RFC 9457. Carries `decision_id` and `policy_version` whenever a policy was involved, so
        a refusal is auditable and reproducible rather than an opaque denial.
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string, examples: [UAI_IDENTITY_REVOKED, UAI_POP_REQUIRED, UAI_CHAIN_CONFLICT] }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        decision_id: { type: string }
        policy_version: { type: string }
        governance_proof: { type: string }
        remediation: { type: string }
        trace_id: { type: string }
...
    VoteStatement:
      $ref: '../schemas/vote-statement.schema.json'
    RevocationDecision:
      $ref: '../schemas/revocation-decision.schema.json'
...